fix: kill the process when Program end

+++ src/index.ts
@@ -61,7 +61,9 @@ function cleanup() {
     }
   }
   for (const subProcess of subProcesses) {
-    subProcess.kill('SIGINT')
+    if (!subProcess.killed) {
+      subProcess.kill('SIGINT')
+    }
   }
 }
